[FIX]: 임시 비밀번호 발급, 댓글 조회 API 수정  (#59)
