+++ src/main/java/com/rom/domain/user/domain/repository/UserRepository.java
@@ -12,6 +12,5 @@ public interface UserRepository extends JpaRepository<User,Long>{
     
     Optional<User> findByEmail(String email);
     Boolean existsByEmail(String email);
-
     Optional<User> findByEmailAndNickname(String email, String nickname);
 }
