+++ src/main/java/com/rom/domain/user/presentation/UserController.java
@@ -2,7 +2,6 @@ package com.rom.domain.user.presentation;
 
 import com.rom.domain.user.application.UserService;
 import com.rom.domain.user.dto.ChangePasswordReq;
-import com.rom.domain.user.dto.SendPasswordReq;
 import com.rom.domain.user.dto.UserDetailRes;
 import com.rom.global.config.security.token.CurrentUser;
 import com.rom.global.config.security.token.UserPrincipal;
@@ -10,7 +9,6 @@ import com.rom.global.payload.ErrorResponse;
 import com.rom.global.payload.Message;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -110,15 +108,4 @@ public class UserController {
         return userService.deleteUser(userPrincipal);
     }
 
-    @Operation(summary = "이메일로 임시 비밀번호 전송")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "이메일로 임시 번호를 전송했습니다.", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = Message.class))}),
-            @ApiResponse(responseCode = "400", description = "이메일로 임시 번호를 전송하지 못했습니다..", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))})
-    })
-    @PostMapping("/me/send")
-    public ResponseEntity<?> sendTemporaryPassword(
-            @Parameter(description = "SendPasswordReq를 참고해주세요.", required = true) @Valid @RequestBody SendPasswordReq sendPasswordReq
-    ){
-        return userService.sendTemporaryPassword(sendPasswordReq);
-    }
 }
